Deduplicate common columns
