+++ packages/desktop/app/components/PatientPrinting/modals/PrintMultipleImagingRequestsSelectionForm.js
@@ -5,28 +5,15 @@ import { useQuery } from '@tanstack/react-query';
 import { Table, useSelectableColumn } from '../../Table';
 import { OuterLabelFieldWrapper } from '../../Field';
 import { ConfirmCancelRow } from '../../ButtonRow';
-import { DateDisplay } from '../../DateDisplay';
 import { useApi } from '../../../api';
 import { Colors } from '../../../constants';
-import { getImagingRequestType } from '../../../utils/getImagingRequestType';
-import { getAreaNote } from '../../../utils/areaNote';
-import { useLocalisation } from '../../../contexts/Localisation';
 
 import { MultipleImagingRequestsPrintoutModal } from './MultipleImagingRequestsPrintoutModal';
-
-const COLUMN_KEYS = {
-  ID: 'id',
-  SELECTED: 'selected',
-  REQUESTED_DATE: 'requestedDate',
-  REQUESTED_BY: 'requestedBy',
-  TYPE: 'imagingType',
-  AREAS: 'areas',
-};
+import { COLUMN_KEYS, FORM_COLUMNS } from './multipleImagingRequestsColumns';
 
 export const PrintMultipleImagingRequestsSelectionForm = React.memo(({ encounter, onClose }) => {
   const [openPrintoutModal, setOpenPrintoutModal] = useState(false);
   const api = useApi();
-  // TODO: make sure endpoint supports query parameters
   const { data: imagingRequestsData, error, isLoading } = useQuery(
     ['imagingRequests', encounter.id],
     async () => {
@@ -53,42 +40,7 @@ export const PrintMultipleImagingRequestsSelectionForm = React.memo(({ encounter
     }
   }, [selectedRows]);
 
-  const { getLocalisation } = useLocalisation();
-  const imagingTypes = getLocalisation('imagingTypes') || {};
-  const columns = [
-    selectableColumn,
-    {
-      key: COLUMN_KEYS.ID,
-      title: 'Request ID',
-      sortable: false,
-    },
-    {
-      key: COLUMN_KEYS.REQUESTED_DATE,
-      title: 'Request date',
-      sortable: false,
-      accessor: ({ requestedDate }) => <DateDisplay date={requestedDate} />,
-    },
-    {
-      key: COLUMN_KEYS.REQUESTED_BY,
-      title: 'Requested by',
-      sortable: false,
-      maxWidth: 300,
-      accessor: ({ requestedBy }) => requestedBy?.displayName || '',
-    },
-    {
-      key: COLUMN_KEYS.TYPE,
-      title: 'Type',
-      sortable: false,
-      maxWidth: 70,
-      accessor: getImagingRequestType(imagingTypes),
-    },
-    {
-      key: COLUMN_KEYS.AREAS,
-      title: 'Areas to be imaged',
-      sortable: false,
-      accessor: getAreaNote,
-    },
-  ];
+  const columns = [selectableColumn, ...FORM_COLUMNS];
 
   return (
     <>
